pollsite/requirements.txt
Django==1.8
django-polymorphic==0.7
wsgiref==0.1.2
